requirements.txt
asgiref==3.8.1
attrs==25.1.0
beautifulsoup4==4.13.3
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
coverage==7.8.0
cryptography==44.0.2
defusedxml==0.7.1
Deprecated==1.2.18
Django==4.2.20
django-bootstrap-v5==1.0.11
django-bootstrap5==25.1
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
drf-spectacular==0.28.0
execnet==2.1.1
factory_boy==3.3.3
Faker==37.1.0
h11==0.14.0
idna==3.10
img2pdf==0.6.0
inflection==0.5.1
iniconfig==2.0.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lxml==5.3.2
MarkupSafe==3.0.2
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.2
pdf2image==1.17.0
pikepdf==9.7.0
pillow==11.2.1
pluggy==1.5.0
psutil==7.0.0
pycparser==2.22
PyJWT==2.9.0
PySocks==1.7.1
pytest==8.3.5
pytest-base-url==2.1.0
pytest-cov==6.1.1
pytest-django==4.10.0
pytest-factoryboy==2.7.0
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-mock==3.14.0
pytest-selenium==4.1.0
pytest-variables==3.1.0
pytest-xdist==3.6.1
python-docx==1.1.2
python-dotenv==1.0.1
python3-openid==3.2.0
PyYAML==6.0.2
referencing==0.36.2
reportlab==4.4.0
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.23.1
selenium==4.31.0
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
tenacity==9.1.2
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
websocket-client==1.8.0
wrapt==1.17.2
wsproto==1.2.0
